TAC-1580, TAC-1581, TAC-1586, TAC-1579, TAC-1587, TAC-1590, TAC-1591, TAC-1592, TAC-1593, changes after testing

+++ aspnet-core/src/TACHYON.Core/Rating/RatingLog.cs
@@ -38,7 +38,7 @@ namespace TACHYON.Rating
         [ForeignKey("FacilityId")]
         public Facility FacilityFk { get; set; }
         public RateType RateType { get; set; }
-        public int Rate { get; set; }
+        public decimal Rate { get; set; }
         public string Note { get; set; }
         public DateTime CreationTime { get; set; }
         /// <summary>
